Adds name server client;
Also:
- Moves demo to test file
