+++ pubspec.yaml
@@ -15,8 +15,11 @@ dependencies:
   http: ">=0.13.3 <1.0.0"
   flutter_bloc: ">=7.0.0 <8.0.0"
   web_socket_channel: ">=2.1.0 <3.0.0"
+  namester:
+    path: packages/namester
 
 dev_dependencies:
+  # pedantic: ^1.9.0
   # test: "<1.17.0" # use the version below 1.17.0 to avoid clashes with flutter_test
   flutter_test:
     sdk: flutter
